Remove check for screen size
It seems agreed upon that a terminal is 80x24 so I'm going to accept this as the general size and not worry about too strictly enforcing width/height constraints. I'm sure there's some kind of standard specifying a terminal screen is 80x24 since the Linux kernel documentation mentions a standard.